# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bluebulb Expense Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  requester: "requester",
  approver_manager: "approver_manager",
  internal_control: "internal_control",
  finance_manager: "finance_manager",
  finance_team: "finance_team",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  department: string;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  message: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  requester: "requester",
  approver_manager: "approver_manager",
  internal_control: "internal_control",
  finance_manager: "finance_manager",
  finance_team: "finance_team",
} as const;

export interface CreateUserRequest {
  name: string;
  email: string;
  password: string;
  role: CreateUserRequestRole;
  department: string;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  requester: "requester",
  approver_manager: "approver_manager",
  internal_control: "internal_control",
  finance_manager: "finance_manager",
  finance_team: "finance_team",
} as const;

export interface UpdateUserRequest {
  name?: string;
  email?: string;
  role?: UpdateUserRequestRole;
  department?: string;
}

export type ExpenseRequestStatus =
  (typeof ExpenseRequestStatus)[keyof typeof ExpenseRequestStatus];

export const ExpenseRequestStatus = {
  draft: "draft",
  pending_manager: "pending_manager",
  pending_internal_control: "pending_internal_control",
  pending_finance_manager: "pending_finance_manager",
  pending_payment: "pending_payment",
  paid: "paid",
  rejected: "rejected",
} as const;

export interface ExpenseRequest {
  id: number;
  expenseName: string;
  expenseType: string;
  purpose: string;
  otherInfo?: string;
  department: string;
  requesterId: number;
  requesterName: string;
  approverId: number;
  approverName: string;
  internalControlId?: number;
  internalControlName?: string;
  financeManagerId?: number;
  financeManagerName?: string;
  requestDate: string;
  dueDate?: string;
  bankName: string;
  accountNumber: string;
  accountName: string;
  amount: number;
  status: ExpenseRequestStatus;
  currentApprovalStep?: string;
  attachmentUrl?: string;
  receiptUrl?: string;
  managerApprovalDate?: string;
  managerRejectionReason?: string;
  internalControlApprovalDate?: string;
  internalControlRejectionReason?: string;
  financeManagerApprovalDate?: string;
  financeManagerRejectionReason?: string;
  paidDate?: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateExpenseRequestStatus =
  (typeof CreateExpenseRequestStatus)[keyof typeof CreateExpenseRequestStatus];

export const CreateExpenseRequestStatus = {
  draft: "draft",
  pending_manager: "pending_manager",
} as const;

export interface CreateExpenseRequest {
  expenseName: string;
  expenseType: string;
  purpose: string;
  otherInfo?: string;
  department: string;
  requesterId: number;
  approverId: number;
  internalControlId?: number;
  financeManagerId?: number;
  requestDate: string;
  dueDate?: string;
  bankName: string;
  accountNumber: string;
  accountName: string;
  amount: number;
  status: CreateExpenseRequestStatus;
  attachmentUrl?: string;
}

export type UpdateExpenseRequestStatus =
  (typeof UpdateExpenseRequestStatus)[keyof typeof UpdateExpenseRequestStatus];

export const UpdateExpenseRequestStatus = {
  draft: "draft",
  pending_manager: "pending_manager",
} as const;

export interface UpdateExpenseRequest {
  expenseName?: string;
  expenseType?: string;
  purpose?: string;
  otherInfo?: string;
  department?: string;
  approverId?: number;
  internalControlId?: number;
  financeManagerId?: number;
  requestDate?: string;
  dueDate?: string;
  bankName?: string;
  accountNumber?: string;
  accountName?: string;
  amount?: number;
  status?: UpdateExpenseRequestStatus;
  attachmentUrl?: string;
}

export type ApprovalActionRequestAction =
  (typeof ApprovalActionRequestAction)[keyof typeof ApprovalActionRequestAction];

export const ApprovalActionRequestAction = {
  approve: "approve",
  reject: "reject",
} as const;

export interface ApprovalActionRequest {
  action: ApprovalActionRequestAction;
  reason?: string;
  approverId: number;
}

export interface MarkPaidRequest {
  receiptUrl?: string;
  paidById: number;
}

export type RetirementRequestStatus =
  (typeof RetirementRequestStatus)[keyof typeof RetirementRequestStatus];

export const RetirementRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface RetirementRequest {
  id: number;
  expenseId: number;
  expenseName: string;
  requesterId: number;
  requesterName: string;
  department: string;
  amount: number;
  actualAmount?: number;
  purpose: string;
  status: RetirementRequestStatus;
  approverId?: number;
  approverName?: string;
  approvalDate?: string;
  rejectionReason?: string;
  attachmentUrl?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateRetirementRequest {
  expenseId: number;
  requesterId: number;
  actualAmount: number;
  purpose: string;
  approverId: number;
  attachmentUrl?: string;
}

export interface DepartmentStat {
  department: string;
  count: number;
  amount: number;
}

export interface MonthStat {
  month: string;
  count: number;
  amount: number;
}

export interface YearStat {
  year: string;
  count: number;
  amount: number;
}

export interface StatusStat {
  status: string;
  count: number;
  amount: number;
}

export interface DashboardStats {
  totalExpenses: number;
  totalAmount: number;
  pendingCount: number;
  approvedCount: number;
  rejectedCount: number;
  paidCount: number;
  byDepartment: DepartmentStat[];
  byMonth: MonthStat[];
  byYear: YearStat[];
  byStatus: StatusStat[];
}

export type GetExpensesParams = {
  requesterId?: number;
  status?: string;
  approverId?: number;
};

export type GetRetirementsParams = {
  requesterId?: number;
  status?: string;
};